Adds delivery resource and example data
Adds delivery resource that is a join between material and users.
I'd like to add multiple references to users as we need a prepper,
deliverer, reviewer, maybe on-point person, and maybe 'owner/creator'
but none of that has been decided yet. Example data also added to
prove the hasmany relationship works and to make testing easier during
development.