Seed database with materials parsed from ICS
Adds code to `db/seeds.rb` to run `scripts/parse-ics.rb` and seed the
database with the materials parsed from `public/cal.ics`. This should
include all the materials that were assigned for 019 (because I used
their calendar file), but we definitely need to check to make sure it's
all there, and that nothing erroneous got through.